botonera_http/requirements.txt
requests==2.31.0
RPi.GPIO==0.7.1
